♻️ refactor: rename package from app.wasm to app.web for unified web target
- Migrate package structure: mikufan.cx.conduit.frontend.app.wasm → mikufan.cx.conduit.frontend.app.web
- Update package declarations in Main.kt and all setup files
- Move lifecycle management utilities to extension functions on LifecycleRegistry
- Remove misleading WASM-specific naming now that code serves unified web target (JS + WASM)
